+++ packages/polaris-viz-core/src/components/LineSeries/components/SparkArea/index.ts
@@ -1,0 +1,1 @@
+export {SparkArea} from './SparkArea';