+++ packages/polaris-viz/src/hooks/ColorVisionA11y/index.ts
@@ -1,5 +1,2 @@
-export {getColorVisionEventAttrs} from './getColorVisionEventAttrs';
 export {useColorVisionEvents} from './useColorVisionEvents';
 export {useWatchColorVisionEvents} from './useWatchColorVisionEvents';
-export {COLOR_VISION_EVENT} from './constants';
-export {getOpacityStylesForActive} from './utilities';